addons/pinniped/tanzu-auth-controller: rename config-controller to tanzu-auth-controller (#2250)
