+++ addons/pinniped/tanzu-auth-controller-manager/hack/ytt/default-package-secret.yaml
