+++ addons/pinniped/tanzu-auth-controller-manager/hack/ytt/schema.yaml
